Move each backup api type into its own file
