+++ data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/BackupDestination.kt
@@ -1,0 +1,7 @@
+package com.thomaskioko.tvmaniac.data.backup.api
+
+public interface BackupDestination {
+    public fun write(contents: String)
+
+    public fun read(): String
+}
