+++ data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/BackupDestinationBuilder.kt
@@ -1,0 +1,7 @@
+package com.thomaskioko.tvmaniac.data.backup.api
+
+public interface BackupDestinationBuilder {
+    public fun build(location: String): BackupDestination
+
+    public fun safetyCopy(): BackupDestination
+}
