+++ data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/BackupFormat.kt
@@ -1,0 +1,9 @@
+package com.thomaskioko.tvmaniac.data.backup.api
+
+public object BackupFormat {
+    public const val VERSION: Int = 1
+    public const val FILE_PREFIX: String = "tvmaniac-backup-"
+    public const val FILE_EXTENSION: String = ".json"
+    public const val MIME_TYPE: String = "application/json"
+    public const val SAFETY_COPY_NAME: String = "tvmaniac-pre-restore.json"
+}
