+++ data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/BackupLocationUnreadableException.kt
@@ -1,0 +1,3 @@
+package com.thomaskioko.tvmaniac.data.backup.api
+
+public class BackupLocationUnreadableException(location: String) : Exception("Cannot open $location")
