+++ data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/RestoreFailure.kt
@@ -1,0 +1,9 @@
+package com.thomaskioko.tvmaniac.data.backup.api
+
+public enum class RestoreFailure {
+    SyncInProgress,
+    ReadFailed,
+    VersionTooNew,
+    SafetyCopyFailed,
+    ImportFailed,
+}
