+++ data/backup/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/data/backup/api/RestoreSummary.kt
@@ -1,0 +1,10 @@
+package com.thomaskioko.tvmaniac.data.backup.api
+
+public data class RestoreSummary(
+    val showCount: Int,
+    val episodeCount: Int,
+    val skippedShows: List<String> = emptyList(),
+    val skippedSeasonRatings: Int = 0,
+    val skippedEpisodeRatings: Int = 0,
+    val rewatchSessionsKept: Int = 0,
+)
